--- a/torchaudio/datasets/vctk.py
+++ b/torchaudio/datasets/vctk.py
@@ -14,7 +14,6 @@ from torchaudio.datasets.utils import (
 
 URL = "https://datashare.is.ed.ac.uk/bitstream/handle/10283/3443/VCTK-Corpus-0.92.zip"
 FOLDER_IN_ARCHIVE = "VCTK-Corpus"
-FOLDER_IN_ARCHIVE_092 = "VCTK-Corpus-0.92"
 _CHECKSUMS = {
     "https://datashare.is.ed.ac.uk/bitstream/handle/10283/3443/VCTK-Corpus-0.92.zip":
     "8a6ba2946b36fcbef0212cad601f4bfa"
@@ -150,25 +149,17 @@ class VCTK_092(Dataset):
     For more information about the dataset visit: https://datashare.is.ed.ac.uk/handle/10283/3443
     """
 
-    _folder_txt = "txt"
-    _folder_audio = "wav48_silence_trimmed"
-    _ext_txt = ".txt"
-    _ext_audio = ".flac"
-    _mic_1 = "mic1"
-    _mic_2 = "mic2"
-
     def __init__(self,
                  root: str,
                  url: str = URL,
-                 download: bool = False) -> None:
+                 download: bool = False,
+                 mic_id: str = 'mic2') -> None:
 
-        # Custom url may contain options like '?sequence=2&isAllowed=y'
-        # Slicing it off to extract only 'VCTK-Corpus-0.92.zip'
-        archive = os.path.join(root, os.path.basename(url)[:20])
+        archive = os.path.join(root, os.path.basename('VCTK-Corpus-0.92.zip'))
 
-        self._path = os.path.join(root, FOLDER_IN_ARCHIVE_092)
-        self._txt_dir = os.path.join(self._path, self._folder_txt)
-        self._audio_dir = os.path.join(self._path, self._folder_audio)
+        self._path = os.path.join(root, 'VCTK-Corpus-0.92')
+        self._txt_dir = os.path.join(self._path, 'txt')
+        self._audio_dir = os.path.join(self._path, 'wav48_silence_trimmed')
 
         if download:
             if not os.path.isdir(self._path):
@@ -183,32 +174,32 @@ class VCTK_092(Dataset):
             )
 
         # Extracting speaker IDs from the folder structure
-        self._speaker_ids = sorted(d for d in os.listdir(self._txt_dir)
-                                   if os.path.isdir(os.path.join(self._txt_dir, d)))
-        self._audio_ids = []
+        self._speaker_ids = sorted(os.listdir(self._txt_dir))
+        self._sample_ids = []
 
         """
         Due to some insufficient data complexity in the 0.92 version of this dataset,
         we start traversing the audio folder structure in accordance with the text folder.
-        As some of the audio files are missing of either ``mic_1`` or ``mic_2`` but the 
+        As some of the audio files are missing of either ``mic_1`` or ``mic_2`` but the
         text is present for the same, we first check for the existence of the audio file
-        before adding it to the ``audio_ids`` list.
+        before adding it to the ``sample_ids`` list.
 
         Once the ``audio_ids`` are loaded into memory we can quickly access the list for
         different parameters required by the user.
         """
         for speaker_id in self._speaker_ids:
             utterance_dir = os.path.join(self._txt_dir, speaker_id)
-            for utterance_file in sorted(f for f in os.listdir(utterance_dir) if f.endswith(self._ext_txt)):
+            for utterance_file in sorted(f for f in os.listdir(utterance_dir) if f.endswith('.txt')):
                 utterance_id = os.path.splitext(utterance_file)[0]
-                audio_path_mic1 = os.path.join(self._audio_dir, speaker_id,
-                                               f'{utterance_id}_{self._mic_1}{self._ext_audio}')
-                audio_path_mic2 = os.path.join(self._audio_dir, speaker_id,
-                                               f'{utterance_id}_{self._mic_2}{self._ext_audio}')
-                if os.path.exists(audio_path_mic1):
-                    self._audio_ids.append(f'{utterance_id}_{self._mic_1}')
-                if os.path.exists(audio_path_mic2):
-                    self._audio_ids.append(f'{utterance_id}_{self._mic_2}')
+                audio_path_mic = os.path.join(self._audio_dir, speaker_id,
+                                              f'{utterance_id}_{mic_id}.flac')
+                if speaker_id == 'p280' and mic_id == 'mic2':
+                    break
+                elif speaker_id == 'p362':
+                    if os.path.isfile(audio_path_mic):
+                        self._sample_ids.append(f'{utterance_id}_{mic_id}')
+                else:
+                    self._sample_ids.append(f'{utterance_id}_{mic_id}')
 
     def load_text(self, file_path) -> str:
         with open(file_path) as file_path:
@@ -217,22 +208,24 @@ class VCTK_092(Dataset):
     def load_audio(self, file_path) -> Tuple[Tensor, int]:
         return torchaudio.load(file_path)
 
-    def load_sample(self, audio_id: str) -> Sample:
-        speaker_id, utterance_id, _ = audio_id.split("_")
-        utterance_path = os.path.join(self._txt_dir, speaker_id, f'{speaker_id}_{utterance_id}{self._ext_txt}')
-        audio_path = os.path.join(self._audio_dir, speaker_id, f'{audio_id}{self._ext_audio}')
+    def load_sample(self, sample_id: str) -> Sample:
+        # Slicing the speaker and utterance IDs
+        speaker_id, utterance_id, _ = sample_id.split("_")
 
-        # Read text
+        utterance_path = os.path.join(self._txt_dir, speaker_id, f'{speaker_id}_{utterance_id}.txt')
+        audio_path = os.path.join(self._audio_dir, speaker_id, f'{sample_id}.flac')
+
+        # Reading text
         utterance = self.load_text(utterance_path)
 
-        # Read FLAC
+        # Reading FLAC
         waveform, sample_rate = self.load_audio(audio_path)
 
         return Sample(waveform, sample_rate, utterance, speaker_id, utterance_id)
 
     def __getitem__(self, n: int) -> Sample:
-        audio_id = self._audio_ids[n]
-        return self.load_sample(audio_id)
+        sample_id = self._sample_ids[n]
+        return self.load_sample(sample_id)
 
     def __len__(self) -> int:
-        return len(self._audio_ids)
+        return len(self._sample_ids)