- Updated dataset URL's
- Updated the zip file checksum
- Fixed URL path parsing for the updated link
- Updated audio folder and audio extension names

The last URL returned a 404 HTTP not found error as the dataset was moved to a new link.

Verified the changes by loading the dataset into dataloader and looping over every file.
Tested by training some epochs.