@Abhi011999 2020-07-24
Hi @mthrok ! Thanks for reviewing this.

I made some changes. Hope these help.