@mthrok 2020-08-11
Hi @Abhi011999 

Sorry for the late reply. We were discussing internal how we should approach this problem.
We decided to keep the original VCTK as is (but raise an error when `download=True` with the message to use a new version.), then add the new version, (say `VCTK_V2`).

Could you move your changes to `VCTK_V2`? You can either copy-paste the original `VCTK` or subclass `VCTK` (if that works and is simpler)