@Abhi011999 2020-08-11
@mthrok An optional boolean argument can be added in the existing `VCTK` class that identifies the old and the new dataset with the default value set to `False`. That can work?