@mthrok 2020-08-11
> @mthrok An optional boolean argument can be added in the existing `VCTK` class that identifies the old and the new dataset with the default value set to `False`. That can work?

It would, but we do not know if there will be another dataset change like this (though chances are small),
and boolean is not future proof.
Also, the previous version is now generally inaccessible, yet downloading the new one takes a bit time.
So rather than mixing the code for dataset that are only useful for those who has cached data and the code 
that general public can use, we concluded it's cleaner to separate them.
Maybe we should call the new dataset `VCTK_092`.