@mthrok 2020-08-12
> @mthrok So the extra comments are not necessary or should I just leave them?

Which one do you mean by "the extra comments" ??


Also, can you put the new class `VCTK_092` into the same module, `vctk.py`?
That way, you do not need to duplicate `load_vctk_item` function, right?