@Abhi011999 2020-08-12
> That way, you do not need to duplicate `load_vctk_item` function, right?

No, some lines are different for the new version.

> Which one do you mean by "the extra comments" ??

I mean the comments I put when I made changes to the original `vctk.py`.

> can you rebase your work?

will do, just want a final confirmation.