@mthrok 2020-08-12
> > That way, you do not need to duplicate `load_vctk_item` function, right?
> 
> No, some lines are different for the new version.

I see. In that case, could you make it to a method?
That way it's easier to allow users to customize the loading behavior if necessary.

> > Which one do you mean by "the extra comments" ??
> 
> I mean the comments I put when I made changes to the original `vctk.py`.

I can no longer see the original commit. Could you describe it?

Also can you put the new class in the same `vctk.py` module?
I think adding a new submodule for a new version of existing dataset might get 
unmanageable at some point if there are many updates.

Once mode thing, can you change the original VCTK class so that it will raise 
an error if `download=True` with an error message that explains the dataset is
no longer available and users should use the new class you are adding.