@mthrok 2020-08-13
> I am a little out of knowledge here but why are the tests failing exactly?
> 
> EDIT: Is my branch `patch-1` in conflict here?

It's not conflict but there was a breaking change in PyTorch core and we added the change for that in #850 .
I see that your branch is based off of #00cc000, which does not have this fix. So if you fetch the latest master branch of `pytorch/audio` and rebase your changes on it, then the tests should pass.