@mthrok 2020-08-14
> I see random tests failing and passing. Seems like I have to pass the "torch" to you now @mthrok 👀

lol nice one. Yeah, unfortunately we are hit by another upstream issue. https://github.com/pytorch/pytorch/pull/43050
CI should be working fine tomorrow.

Meanwhile there are couple of improvements on the current code base. If you have time, could you give some thoughts? This will make the implementation of `VCTK_092` much better.

1. Making `load_vctk_092_item` function into a method of `VCTK_092`.
    Putting `load_vctk_092_item` as a plain function makes it difficult to customize the behavior on user code. If a user wants to change a bit of behavior, the user has to re-write the whole class. 
2. Make the return value typed.
    In general, when returning multiple objects from a function, it a good idea to introduce micro protocol with `namedtuple`. This way the access to the individual item gets cleaner and it becomes future-prof (adding a new item to be returned is backward compatible.) Check out [my example here](https://github.com/mthrok/audio/blob/wsj0-mix-dataset/torchaudio/datasets/wsj0mix.py#L10).
3. Wrap `torchaudio.load` as another method of `VCTK_092`.
    Similar to 1, this will allow user to customize the loading behavior, like disabling normalization or swapping channels. See example [here](https://github.com/mthrok/audio/blob/c103e39ee4a811f52d26d4f30a0e04e1d52f81ce/torchaudio/datasets/wsj0mix.py#L22-L23)
4. Get rid of `walk_files` function.
    Once the root directory is located, we have the idea of expected directory structure, so there is no need to traverse directories as if it's unknown. `walk_files` function can do harm by traversing unrelated directories, but it does not bring any benefit.

    ```
    speaker_ids = sorted(d for d in os.listdir(txt_dir) if os.path.isdir(os.path.join(txt_dir, d)))
    for speaker_id in speaker_ids:
        utterance_dir = os.path.join(txt_dir, speaker_id)
        speech_dir = os.path.join(audio_dir, speaker_id)
        for utterance_file in sorted(f for f in os.listdir(utterance_dir) if f.endswith("txt")):
            utterance_id = os.path.splitext(utterance_file)[0]
            utterance_path = os.path.join(utterance_dir, utterance_id)
            audio_path = os.path.join(audio_dir, f'{utterance_id}_{mic_id}.{self.audio_ext}'
    ```
    
    This way, you do not need attributes like `_except_folder`. and the directory traversal is minimum.