@Abhi011999 2020-08-15
@mthrok 

So I created a test script for the new implementation based on your above approach **without** `walk_files` and guess what?

Half of the `mic1` audio files have been removed from `p362`, but there `occurrence` is present in the `txt` folder. Below is the full list with IDs -
https://pastebin.com/p02Fyby7

My take here is that after the total path is calculated of every possible audio file, it performs a validation check if the file exists or not using `os.path.exists` method and then add it to a list.
What do you think?

Note - Everything else has been taken care of, only this remains.