@Abhi011999 2020-08-17
Hi, @mthrok. As I have updated with the required changes, I also made some performance comparisons with and without the `walk_files` implementation and here are some insights -

- With `walk_files`
  - `__init__()` takes 0.5 seconds approx. to walk over files.
  - Retrieving one sample takes about 0.05 seconds approx.
  - Iterating over the entire dataset takes about 18 min 26 secs.
  - It can wrongly walk over any file in any nested directory with the same extensions.
  - We need to explicitly specify which folders or files to omit.

- Without `walk_files`
  - `__init__()` takes 7 seconds approx. to walk over files.
  - Retrieving one sample takes about 0.048 seconds approx.
  - Iterating over the entire dataset takes about 18 min 26 secs.
  - This implementation is future proof for additional upcoming updates.

The conclusion that we can draw from the above stats is that we can keep the new implementation without `walk_files` as its future proof and further relevant changes can be made easily in accordance with the upcoming updates for the dataset.

That's my take. Would love to hear what you think.